requirements.txt
aiogram==2.25.1
asyncpg==0.28.0
python-dotenv==1.0.0